Move debug render tree bookkeeping onto the tree object
The component and modifier handlers carried the code that feeds the
debug render tree inside `if (vm.env.debugRenderTree !== undefined)`
blocks: args capture, custom lifecycle nodes, destructor registration,
and two updating opcodes. That is about 5 kB of source in every bundle
even when the tree is left out.

Those blocks are now methods on `DebugRenderTreeImpl`
(`componentDidGetSelf`, `componentDidRenderLayout`, `modifierDidAdd`,
`remoteElementDidPush`), and the handlers call them through
`debugTree(vm.env)?.`. A build that opts out of the tree drops the
bookkeeping with it. The runtime only ever receives `DebugRenderTreeImpl`,
which is why `debugTree()` can narrow the host-facing type.
