+++ packages/@glimmer/runtime/lib/debug-render-tree.ts
@@ -1,17 +1,35 @@
 import { DEBUG } from '@glimmer/env';
 import type {
   Bounds,
+  CapturedArguments,
   CapturedRenderNode,
+  CompilableProgram,
   ComponentDefinition,
+  ComponentInstance,
   DebugRenderTree,
+  Environment,
+  ModifierInstance,
   Nullable,
   RenderNode,
+  UpdatingOpcode,
 } from '@glimmer/interfaces';
+import type { Reference } from '@glimmer/reference/lib/reference';
 import { expect } from '@glimmer/debug-util/lib/platform-utils';
 import { assign } from '@glimmer/util/lib/object-utils';
 import { StackImpl as Stack } from '@glimmer/util/lib/collections';
 
-import { reifyArgsDebug } from './vm/arguments';
+import type { VM } from './vm/append';
+import type { UpdatingVM } from './vm';
+
+import { registerDestructor } from '@glimmer/destroyable';
+import { managerHasCapability } from '@glimmer/manager/lib/util/capabilities';
+import { valueForRef } from '@glimmer/reference/lib/reference';
+import { InternalComponentCapabilities } from '@glimmer/vm/lib/flags';
+import assert from '@glimmer/debug-util/lib/assert';
+
+import { ConcreteBounds } from './bounds';
+import { hasCustomDebugRenderTreeLifecycle } from './component/interfaces';
+import { createCapturedArgs, EMPTY_ARGS, reifyArgsDebug } from './vm/arguments';
 
 interface InternalRenderNode<T extends object> extends RenderNode {
   bounds: Nullable<Bounds>;
@@ -100,6 +118,158 @@ export default class DebugRenderTreeImpl<
 
   commit(): void {
     this.reset();
+  }
+
+  componentDidGetSelf(
+    vm: VM,
+    instance: ComponentInstance,
+    selfRef: Reference,
+    namesHandle: number
+  ): void {
+    let { definition, manager, state } = instance;
+
+    let args: CapturedArguments;
+
+    if (vm.stack.peek() === vm.args) {
+      args = vm.args.capture();
+    } else {
+      let names = vm.constants.getArray<string>(namesHandle);
+      vm.args.setup(vm.stack, names, [], 0, true);
+      args = vm.args.capture();
+    }
+
+    let compilable: CompilableProgram | null = definition.compilable;
+
+    if (compilable === null) {
+      assert(
+        managerHasCapability(
+          manager,
+          instance.capabilities,
+          InternalComponentCapabilities.dynamicLayout
+        ),
+        'BUG: No template was found for this component, and the component did not have the dynamic layout capability'
+      );
+
+      let resolver = vm.context.resolver;
+      compilable = resolver === null ? null : manager.getDynamicLayout(state, resolver);
+    }
+
+    // For tearing down the debugRenderTree
+    vm.associateDestroyable(instance);
+
+    if (hasCustomDebugRenderTreeLifecycle(manager)) {
+      let nodes = manager.getDebugCustomRenderTree(instance.definition.state, instance.state, args);
+
+      nodes.forEach((node) => {
+        let { bucket } = node;
+        this.create(bucket as TBucket, node);
+
+        registerDestructor(instance, () => {
+          this.willDestroy(bucket as TBucket);
+        });
+
+        vm.updateWith(new DebugRenderTreeUpdateOpcode(bucket));
+      });
+    } else {
+      let name = getDebugName(definition, manager);
+
+      this.create(instance as unknown as TBucket, {
+        type: 'component',
+        name,
+        args,
+        instance: valueForRef(selfRef),
+      });
+
+      registerDestructor(instance, () => {
+        this.willDestroy(instance as unknown as TBucket);
+      });
+
+      vm.updateWith(new DebugRenderTreeUpdateOpcode(instance));
+    }
+  }
+
+  componentDidRenderLayout(vm: VM, instance: ComponentInstance, bounds: Bounds): void {
+    let { manager, state } = instance;
+
+    if (hasCustomDebugRenderTreeLifecycle(manager)) {
+      let nodes = manager.getDebugCustomRenderTree(instance.definition.state, state, EMPTY_ARGS);
+
+      nodes.reverse().forEach((node) => {
+        let { bucket } = node;
+
+        this.didRender(bucket as TBucket, bounds);
+
+        vm.updateWith(new DebugRenderTreeDidRenderOpcode(bucket, bounds));
+      });
+    } else {
+      this.didRender(instance as unknown as TBucket, bounds);
+
+      vm.updateWith(new DebugRenderTreeDidRenderOpcode(instance, bounds));
+    }
+  }
+
+  modifierDidAdd(vm: VM, modifier: ModifierInstance, capturedArgs: CapturedArguments): void {
+    const { manager, definition, state } = modifier;
+
+    // TODO: we need a stable object for the debugRenderTree as the key, add support for
+    // the case where the state is a primitive, or if in practice we always have/require
+    // an object, then change the internal types to reflect that
+    if (state === null || (typeof state !== 'object' && typeof state !== 'function')) {
+      return;
+    }
+
+    let { element, constructing } = vm.tree();
+    let name = definition.resolvedName ?? manager.getDebugName(definition.state);
+    let instance = manager.getDebugInstance(state);
+
+    assert(constructing, `Expected a constructing element in addModifier`);
+
+    let bounds = new ConcreteBounds(element, constructing, constructing);
+
+    this.create(state as TBucket, {
+      type: 'modifier',
+      name,
+      args: capturedArgs,
+      instance,
+    });
+
+    this.didRender(state as TBucket, bounds);
+
+    // For tearing down the debugRenderTree
+    vm.associateDestroyable(state);
+
+    vm.updateWith(new DebugRenderTreeUpdateOpcode(state));
+    vm.updateWith(new DebugRenderTreeDidRenderOpcode(state, bounds));
+
+    registerDestructor(state, () => {
+      this.willDestroy(state as TBucket);
+    });
+  }
+
+  remoteElementDidPush(
+    block: object,
+    elementRef: Reference,
+    insertBeforeRef: Reference,
+    insertBefore: unknown
+  ): void {
+    // Note that there is nothing to update – when the args for an
+    // {{#in-element}} changes it gets torn down and a new one is
+    // re-created/rendered in its place (see the `Assert`s above)
+    let args = createCapturedArgs(
+      insertBefore === undefined ? {} : { insertBefore: insertBeforeRef },
+      [elementRef]
+    );
+
+    this.create(block as TBucket, {
+      type: 'keyword',
+      name: 'in-element',
+      args,
+      instance: null,
+    });
+
+    registerDestructor(block, () => {
+      this.willDestroy(block as TBucket);
+    });
   }
 
   capture(): CapturedRenderNode[] {
@@ -203,4 +373,54 @@ export function getDebugName(
   manager = definition.manager
 ): string {
   return definition.resolvedName ?? definition.debugName ?? manager.getDebugName(definition.state);
+}
+
+/**
+ * The bookkeeping the VM does on behalf of the debug render tree. It lives
+ * on the tree object rather than in the opcode handlers so that a build
+ * which leaves the tree out also leaves this out.
+ */
+export interface RuntimeDebugRenderTree extends DebugRenderTree {
+  componentDidGetSelf(
+    vm: VM,
+    instance: ComponentInstance,
+    selfRef: Reference,
+    namesHandle: number
+  ): void;
+  componentDidRenderLayout(vm: VM, instance: ComponentInstance, bounds: Bounds): void;
+  modifierDidAdd(vm: VM, modifier: ModifierInstance, args: CapturedArguments): void;
+  remoteElementDidPush(
+    block: object,
+    elementRef: Reference,
+    insertBeforeRef: Reference,
+    insertBefore: unknown
+  ): void;
+}
+
+/**
+ * The environment types its tree as the host-facing `DebugRenderTree`. The
+ * runtime only ever receives `DebugRenderTreeImpl`, which also carries the
+ * VM hooks.
+ */
+export function debugTree(env: Environment): RuntimeDebugRenderTree | undefined {
+  return env.debugRenderTree as RuntimeDebugRenderTree | undefined;
+}
+
+export class DebugRenderTreeUpdateOpcode implements UpdatingOpcode {
+  constructor(private bucket: object) {}
+
+  evaluate(vm: UpdatingVM) {
+    vm.env.debugRenderTree?.update(this.bucket);
+  }
+}
+
+export class DebugRenderTreeDidRenderOpcode implements UpdatingOpcode {
+  constructor(
+    private bucket: object,
+    private bounds: Bounds
+  ) {}
+
+  evaluate(vm: UpdatingVM) {
+    vm.env.debugRenderTree?.didRender(this.bucket, this.bounds);
+  }
 }
