+++ tests/node-vitest/tree-shakability.test.js
@@ -32,6 +32,7 @@ it('[dev] source modules with side effects', () => {
       "@ember/-internals/runtime/lib/ext/rsvp.js",
       "@ember/-internals/runtime/lib/mixins/-proxy.js",
       "@ember/array/index.js",
+      "@ember/array/lib/iterator-extension.js",
       "@ember/debug/index.js",
       "@ember/runloop/index.js",
       "@glimmer/compiler/lib/wire-format-module.js",
@@ -59,6 +60,7 @@ it('[prod] source modules with side effects', () => {
       "@ember/-internals/runtime/lib/ext/rsvp.js",
       "@ember/-internals/runtime/lib/mixins/-proxy.js",
       "@ember/array/index.js",
+      "@ember/array/lib/iterator-extension.js",
       "@ember/runloop/index.js",
       "@glimmer/compiler/lib/wire-format-module.js",
       "@glimmer/validator/index.js",
